Перераспределяет код по модулям + дорабатывает функцию проверки времени встречи
