+++ index.html
@@ -236,7 +236,7 @@
   </template>
   <!-- подключение файлов JS -->
   <script src="js/main.js" type="module"></script>
-  <script src="jsfunctions.js"></script>
+  <script src="js/functions.js"></script>
 </body>
 
 </html>
